Bug fixes on profile picture uploads
