+++ sphinx/screens-detail/chat-detail/tribe-detail/build.gradle
@@ -38,9 +38,11 @@ dependencies {
     implementation project(path: ':sphinx:application:common:menus:menu-bottom-tribe')
     implementation project(path: ':sphinx:application:common:menus:menu-bottom-tribe-profile-pic')
     implementation project(path: ':sphinx:application:common:wrappers:wrapper-common')
+    implementation project(path: ':sphinx:application:common:wrappers:wrapper-message-media')
     implementation project(path: ':sphinx:application:data:concepts:concept-image-loader')
     implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-chat')
     implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-contact')
+    implementation project(path: ':sphinx:application:network:concepts:queries:concept-network-query-chat')
     implementation project(path: ':sphinx:screens-detail:camera:camera-view-model-coordinator')
     implementation project(path: ':sphinx:screens-detail:common:detail-resources')
     implementation project(path: ':sphinx:screens-detail:podcast-player:podcast-player')
