+++ sphinx/screens-detail/chat-detail/tribe-detail/src/main/java/chat/sphinx/tribe_detail/ui/TribeDetailViewState.kt
@@ -13,4 +13,6 @@ internal sealed class TribeDetailViewState: ViewState<TribeDetailViewState>() {
         val accountOwner: Contact,
         val podcast: Podcast?
     ): TribeDetailViewState()
+
+    object TribeProfileUpdating: TribeDetailViewState()
 }
